introduce some more rigorous component structuring
