+++ frontend/src/features/export/export.tsx
@@ -8,19 +8,17 @@ import { ExportXML } from "@/features/export/exportXML";
 
 export function Export() {
   return (
-    <>
-      <Dropdown>
-        <div className="d-grid gap-0">
-          <Dropdown.Toggle variant="success" id="dropdown-basic">
-            <RightPaddedIcon bootstrapIconName="cloud-arrow-down" /> Download
-          </Dropdown.Toggle>
-        </div>
-        <Dropdown.Menu>
-          <ExportDecklist />
-          <ExportXML />
-          <ExportImages />
-        </Dropdown.Menu>
-      </Dropdown>
-    </>
+    <Dropdown>
+      <div className="d-grid gap-0">
+        <Dropdown.Toggle variant="success" id="dropdown-basic">
+          <RightPaddedIcon bootstrapIconName="cloud-arrow-down" /> Download
+        </Dropdown.Toggle>
+      </div>
+      <Dropdown.Menu>
+        <ExportDecklist />
+        <ExportXML />
+        <ExportImages />
+      </Dropdown.Menu>
+    </Dropdown>
   );
 }
